Woo! Progress!
So, the down side is that I wasn't able to get the file read/writing working in a way that was acceptable (more on that later),
but I was able to split the memory graph and cpu graph up into separate threads in a way that runs at an acceptable speed.

I had tried splitting up the reads of the cpu and memory data files and the UI update into different threads. Unfortunately that really
slowed things down, and unfortunately I wasn't able to find a way to speed the process up (yet). If I'm ever able to figure that out I'm probably
going to split the IO and the UI out into different threads, but for now I'm happy