Finally got the ui/io threads working.
Made a very unsafe queue implementation. Don't care. Fight me.
Made all of my mutexes and condition variables global in threads.h/c. Don't care. Fight me.

I made the queue store void pointers so that I didn't have to have two different implementations,
one for the CPU and one for Memory.

I made my mutexes and condition variables global because I'm honestly tired of having a bunch of
gigantic function signatures. Perhaps I could remove the extern keyword and add some sort of
getter to the file to encapsulate it, but idk. Doesn't really matter to me.

Really need to work on finding a better way to call the io thread when needed. Currently
the graph function sleeps for a bit, and the io thread does a sleep as well. Not very good imo.

Also:
I realized that if I just constantly let the io thread pull from /proc/stat and /proc/meminfo
it takes a while for the graph to catch up, so what you'd see on the graph would be from 20
seconds ago. So now I don't queue any more stat updates if either the cpu queue or memory queue
has more than 4 items. This keeps the graph up to date. However, this does tie in with the
sleeps in the ui thread graphing function and the io thread. I'm really going to have to
think about this issue for a while. Maybe it's fine how it is? I mean, it works, but I just
think something about it is gross and isn't obvious at a glance why I'm doing that.
Need to find something more simple and with less voodoo.