Some really nasty stuff going on here. But it's all basically proof of concept.
I'm using a queue for the stats array (even though there will only ever be 1 item) just because it already
exists. I just wanted to get it working. I'm going to go back next and remove the timedPeek and timedDequeue
functions from SHARED_QUEUE. But on the brightside now I'm fetching processes and printing them to a new
window thread. But man is this some nasty tasting spaghetti.

I regret nothing... And everything.