Moved PROC_STATS into a global variable to fix crash.
Usually around the 10 minute mark (sometimes before, sometimes after) the program
would segfault in the UI thread when reading process stats. The error would be
something like "Cannot read data at location 0x****". For some reason putting
this value in a global variable fixed the issue.

I'm now realizing, however, that I may end up having the same issue with passing
the procArena into the UI thread. This was added for displaying the amount of
regions alloc'd when running a debug build. Be sure to remove this.