Kinda got side tracked dealing with valgrind, which kinda ruined the whole evening.
On the bright side I fixed the leak happening in the thread safe queue. I tried using my arenas instead of
malloc, but I had some issues, and I think for now keeping the malloc is fine. Not difficult to manage these
mallocs, just have to make sure I free the remaining items in the queue in my cleanup function.