Added a bunch of logic for CPU percentage by process.
Very very gross and behaves wonky. Tacking the previous and current list of processes
on every iteration of the UI loop. But since the io only fetches the processes every
2 seconds, all of my CPU calculations return to 0. And for that split second there
are CPU percentages displayed, the value displayed is some insane value that's gigantic.
So not only do I need to clean this nasty stuff up, I also need to fix the CPU percentage
calculation.