The CPU percentage is now printing. It's accurate too (as far as I can tell).
Tried adding some r_head_free's to clean up stale CPU and Proc data, but it's causing a segfault.
Could be because I need to protect them with a mutex? The only one I was able to dig into was
the one for process data. Might have to look into that a little more.

Also, I realized long ago that I'm doing all of this on the error handling branch, but I don't really care
anymore. I'm a broken man.