A few minor changes.
Was having another crash after super long periods that was taking
place in the get_cpu_time_now, or whatever it's called, inline
function. The crash was happening in fgets and had something to
do with an _IO_Lock. Then I realized I wasn't closing the file
or making sure that FILE *f wasn't NULL. Makes sense as to why I was
getting this io lock error. Pretty sure that fixed everything, and I
ran the app for a really long period without a segfault, but it's still
really hard to tell.