Dude, I cannot tell you why this was so hard for me. The possible constant shifting of size of the process list made this difficult. The only reason I can think as to why this was so difficult to me is either:
A: I'm a brainlet

or

B: I'm over working on this.

Probably a bit of both. I realized I was having trouble actually focusing on this and was
constantly looking at my phone instead of thinking about the issue. This shouldn't be that
difficult of an issue to solve. How the mighty have fallen.

Oh well, it works mostly how I wanted it to and I think I've covered all of the edge cases.