Okay, I actually think I fixed the menu bug for real this time.
Added the const keyword a lot of places. Removed some magic numbers. There was also probably
some moving around of code. Honestly I can't remember though.

I also noticed that the program was still crashing because of me freeing the head of the
memory arena. I can't for the life of me tell why freeing the CPU or Process arena heads
has no issues (especially when I keep around two records at a time) but for some reason
the memory arena can't handle it. I remedied this issue, or at least I think I did, by
ignoring the root of the problem completely and extended the minimum amount of regions
from 5 to 9. No real reason for the jump of four regions. Probably should look at this
again in the future, but for right now idgaf.