I added the load average, which led to me getting carried away.
Now I'm displaying a processes stats when I hit enter on the process. It replaces the content
of the process list window with the stats for the selected process. When the user hits the 'b'
key the process list is displayed again. Pretty cool stuff imo. Still gotta add more stats and
clean the whole thing up. Also: look into some sort of intermediate layer between threads that
can handle the locking/unlocking of mutexes and waiting/signaling of condition variables.