Fixing issues (#20)
* Fixed the graph point/header color issue. Previously the graph point and header color
were just whatever was set for the CPU graph. The memory graph was ignored.

* Addressed the graph color issue and graph pause issue.
I tried to fix the CPU percentage issue for processes, I even adopted, htop's
calculation for CPU usage, but it didn't help. I'm probably missing something
that they did, but I'm kinda tired of messing with it right now. It seems pretty
accurate out side of the > 100% scenarios, so I'm happy with it.