Okay,
Implemented the ability to have one, two or three stat screens visible. The order that
screens are passed in as command line arguments determines the order in which
they're displayed. For example: if I call 'mtop -mc' then the memory and cpu stat
screens will be displayed with memory on top of cpu. 'mtop -c' would display just the
cpu screen and so on. Doesn't work currently if I pass 3 screens in. If I do that it'll
just fall back to the default layout. Need to do that. Also think it would be a good idea
to not fetch stats if it's corresponding screen is not visible. I tried aftter getting this
taken care of, but a lot of shit broke and I didn't want to deal with it right now.