Now you can determine the order of the three window layout.
Also fixed a little bit of undefined behavior where I was initializing an enum array with
-1 when the lowest enum value was 0. I don't think I was using it though, so it may not
have mattered. I don't remember.

Also want to modify this so you can choose a vertical or horizontal layout. Layout is
horizontal currently. Should be pretty easy I think. But that's a problem for another time.