Added a vertical layout. This layout stacks on the left.
I need to work on the command line arguments. Currently it'll take 'mtop -v' for stack left,
'mtop -v left' for the same and 'mtop -v right' for stack right. I feel like I need to take
a position for each argument. But I think that'll also be a lot more complex. Maybe when the
stack is left the first two window arguments populate the first half of the screen
and the third window argument populates the second half. And when the stack is right the first
window argument populates the left half of the screen and the second and third window arguments
populate the right half. I'll also need a way to do this with the horizontal layout.