View Selection (#21)
* Okay,

Implemented the ability to have one, two or three stat screens visible. The order that
screens are passed in as command line arguments determines the order in which
they're displayed. For example: if I call 'mtop -mc' then the memory and cpu stat
screens will be displayed with memory on top of cpu. 'mtop -c' would display just the
cpu screen and so on. Doesn't work currently if I pass 3 screens in. If I do that it'll
just fall back to the default layout. Need to do that. Also think it would be a good idea
to not fetch stats if it's corresponding screen is not visible. I tried aftter getting this
taken care of, but a lot of shit broke and I didn't want to deal with it right now.

* Now you can determine the order of the three window layout.

Also fixed a little bit of undefined behavior where I was initializing an enum array with
-1 when the lowest enum value was 0. I don't think I was using it though, so it may not
have mattered. I don't remember.

Also want to modify this so you can choose a vertical or horizontal layout. Layout is
horizontal currently. Should be pretty easy I think. But that's a problem for another time.

* Added a vertical layout. This layout stacks on the left.

I need to work on the command line arguments. Currently it'll take 'mtop -v' for stack left,
'mtop -v left' for the same and 'mtop -v right' for stack right. I feel like I need to take
a position for each argument. But I think that'll also be a lot more complex. Maybe when the
stack is left the first two window arguments populate the first half of the screen
and the third window argument populates the second half. And when the stack is right the first
window argument populates the left half of the screen and the second and third window arguments
populate the right half. I'll also need a way to do this with the horizontal layout.

* Okay, got a stack right working.

Next order of business will be implementing the duo vertical layout,
which should be pretty easy. And then implementing the horizontal
layout. But before I do either of those things I might make the dimension
function(s) a little more general. There are too many of them and I'm
repeating way too much code. Looks retarded, is retarded.

* Added duo vertical layout. Still need a way of cleaning up all of these
dimension functions and compacting them, but I don't have time for that now

* I can't remember everything I added. I'm in the early stages of condensing
some of this stuff. Not sure if the table idea is stupid or not. Might have to  wait until I'm merging
everything together to know for sure.

* You know, I'm not really certain this makes it any better. I added three functions:
- One for a full layout
- One for a duo layout
- One for a single layout
They all use the padding table, and the full layout function can handle all layouts and orientations, and the duo layout
can handle both orientations. I might try to just squeeze them into one function, but that might be indention hell and
I don't want that

* Condensed window set up down to one function.
I think I like it more than what I had before, but it is pretty chunky.
Also added the quarters top layout. Next I need to make windows appear
from left to right, top to bottom in relation to the order you pass
command line arguments in.

* Added window titles

* Just a lil clean up

* All layouts are now displayed from
left to right, top to bottom based off of the order they were
passed in as command line arguments.

* You can now use = with flags. Like --vertical=right.
Or you can use --vertical right. doesn't matter.

* Where tf did this come from?

* Realigning