Added actions for processes and for input.
Initially I had (more or less) what lines 162-196 from the UI thread
were doing in the action function. But I realized that read_input, and
I think a few other places, were using the view data, which kinda killed that.
However, I think I can go back to that but track selected pid in the list state,
find that in the current processes, then kill or fetch process stats. We'll
see, but the view data stuff is gross currently