Got both the process and process info views pretty much were I want them.
The only thing I need to do now is work on the display of the info. It's not immediate. So the time
between when you hit 'Enter' and when you see the correct stats could take a second or two. I may
just have to fetch the procesess info ever iteration of the IO loop, which isn't ideal, but it's
not that big of a deal.