Idk how it came to this (#22)
* Initial stage of refactoring the way data is shared between threads.

Now I'm using tasks. I'll construct the data on the IO side, create a task that contains action function pointers
which will be carried out on the UI side. This is a major refactor and this is just a POC

* Got the memory task stuff handled. Now time to move on to the process list
and process info tasks. I have a feeling this is going to be rough.

* Added actions for processes and for input.

Initially I had (more or less) what lines 162-196 from the UI thread
were doing in the action function. But I realized that read_input, and
I think a few other places, were using the view data, which kinda killed that.
However, I think I can go back to that but track selected pid in the list state,
find that in the current processes, then kill or fetch process stats. We'll
see, but the view data stuff is gross currently

* Got both the process and process info views pretty much were I want them.

The only thing I need to do now is work on the display of the info. It's not immediate. So the time
between when you hit 'Enter' and when you see the correct stats could take a second or two. I may
just have to fetch the procesess info ever iteration of the IO loop, which isn't ideal, but it's
not that big of a deal.

* Things are in a pretty good state.

A few things I need to take care of:

1. The option window isn't clearing. It's only being put behind the other windows.
   Clearing the window and refreshing in the input handler does get rid of this stuff,
   but honestly it's worse.
2. Still having that bizarre issue where the windows are overlapping the footer when
   the height of the window is even.
3. Need to handle window resize signal.
4. I think I can still improve the way tasks are handled. I might remove the whole
   group concept. Add cleanup functions to tasks, maybe give each task it's own
   arena and make it responsible for that arena's cleanup? Don't fetch new data
   for a stat until it's previous task has been cleaned up? Lots of ideas.

There's probably more here, but that's all I can think of for now.

* Just some small changes. Handling window resize now. Not super happy with it, but it's not awful.

* Fixed the segfault issue with listState on resize. (I think)

* Added a refresh task

* Man, there's a lot of changes. I have no clue what all is here

* Just removing unused stuff

* Removing unused header